- Removed TOA (https://issues.redhat.com/browse/JGRP-2473) - Removed Whiteboard - Removed DrawApplet - Spring cleaning (IntelliJ analysis)

+++ src/org/jgroups/demos/LockServiceDemo.java
@@ -112,7 +112,7 @@ public class LockServiceDemo implements LockNotification {
                 lock_names=parseLockNames(line.substring("trylock".length()).trim());
 
                 String tmp=lock_names.get(lock_names.size() -1);
-                long timeout=(long)-1;
+                long timeout=-1;
                 try {
                     timeout=Long.parseLong(tmp);
                     lock_names.remove(lock_names.size() -1);
